feat: create entities controllers
create:
- App/src/controller/ArticleController
- App/src/controller/CategoryController
- App/src/controller/CommentController
- App/src/controller/LogController

issue #74
