+++ App/src/controller/ArticleController.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\src\controller;
+
+class ArticleController extends Controller
+{
+
+}
