+++ App/src/controller/BackController.php
@@ -2,7 +2,6 @@
 
 namespace App\src\controller;
 
-use App\src\blogFram\Image;
 use App\src\blogFram\Parameter;
 use App\src\blogFram\Search;
 
@@ -12,21 +11,43 @@ use App\src\blogFram\Search;
 class BackController extends Controller
 {    
     /**
+     * @var ArticleController
+     */
+    protected $articleController;
+
+    /**
+     * @var CategoryController
+     */
+    protected $categoryController;
+
+    /**
+     * @var CommentController
+     */
+    protected $commentController;
+
+    /**
+     * @var UserController
+     */
+    protected $userController;
+
+    /**
      * @var string
      */
     private $controller = 'back';
     
     /**
-     * Logout user
+     * Construct BackController
      *
      * @return void
      */
-    public function logout()
+    public function __construct()
     {
-        $this->session->stop();
-        $this->session->start();
-        $this->alert->addSuccess("A bientôt!");
-        header("Location: ".URL."accueil");
+        parent::__construct();  
+        $this->articleController = new ArticleController;
+        $this->categoryController = new CategoryController;
+        $this->commentController = new CommentController;
+        $this->userController = new UserController;
+        $this->logController = new LogController;
     }
     
     /**
@@ -43,7 +64,8 @@ class BackController extends Controller
             exit;
         }
         if($post->get('submit')) {
-            $this->updateAccount($post);
+            echo 'coucou';
+            $this->userController->updateAccount($post);
         }
         return $this->view->render($this->controller, 'profile', [
             'user' => $this->userDAO->getUser($this->session->get('id'))
@@ -77,147 +99,6 @@ class BackController extends Controller
     }
     
     /**
-     * Update user account by user
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function updateAccount(Parameter $post)
-    {
-        if($post->get('submit')) {
-            if($post->get('delete')) {
-                $this->updateAccountDelete($post);
-            } elseif($post->get('password')) {
-                $this->updateAccountPassword($post);
-            } elseif($post->get('email')) {
-                $this->updateAccountEmail($post);
-            } elseif($post->get('avatar')) {
-                $this->updateAccountAvatar($post);
-            } elseif($post->get('newsletter')) {
-                $this->updateAccountNewsletter($post);
-            }         
-        }
-    }
-    
-    /**
-     * Update user password by user
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function updateAccountPassword(Parameter $post)
-    {
-        if($post->get('password') AND $post->get('passwordConfirm')) {
-            $validate = $this->validation->validateInput('user', $post);
-            if($validate) {
-                $passwordHash = password_hash($post->get('password'), PASSWORD_BCRYPT);
-                if($this->userDAO->updateUser($this->session->get('id'), 'password', $passwordHash)) {
-                    $this->alert->addSuccess("Votre mot de passe a bien été modifié.");
-                } else {
-                    $this->alert->addError("Votre mot de passe n'a pas pu être modifié.");
-                }
-            } 
-        }
-    }
-    
-    /**
-     * Update user email by user
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function updateAccountEmail(Parameter $post)
-    {
-        if($post->get('email')) {
-            $validate = $this->validation->validateInput('user', $post);
-            if($validate) {
-                if($this->userDAO->updateUser($this->session->get('id'), 'email', $post->get('email'))) {
-                    $this->alert->addSuccess("Votre adresse mail a bien été modifiée.");
-                } else {
-                    $this->alert->addError("Votre email n'a pas pu être modifiée.");
-                }
-            }
-        }
-    }
-    
-    /**
-     * update user avatar by user
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function updateAccountAvatar(Parameter $post) 
-    {
-        if($post->get('avatar')) {
-            $image = new Image('avatar', $_FILES['avatar'], $this->session->get('id'));
-            if($image->checkImage('avatar', $_FILES['avatar'], $image::TARGET_AVATAR)) {
-                if($image->upload($_FILES['avatar'])) {
-                    if($this->userDAO->updateUser($this->session->get('id'), 'filename', $image->getFilename())) {
-                        $this->alert->addSuccess("Votre avatar a bien été modifié.");                        
-                    } else {
-                        $this->alert->addError("Votre avatar n'a pas pu être modifié.");
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Update user newsletter subscription by user
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function updateAccountNewsletter(Parameter $post) 
-    {
-        if($post->get('newsletter')) {
-            if($post->get('newsletter') === 'on') {
-                if($this->userDAO->updateUser($this->session->get('id'), 'newsletter', 1)) {
-                    $this->alert->addSuccess("Merci de vous être abonné à la newsletter.");
-                } else {
-                    $this->alert->addError("Une erreur a empêché votre abonnement à la newsletter.");
-                }
-            } elseif($post->get('newsletter') === 'off') {
-                if($this->userDAO->updateUser($this->session->get('id'), 'newsletter', 0)) {
-                    $this->alert->addSuccess("Vous êtes bien désabonné de la newsletter.");
-                } else {
-                    $this->alert->addError("Une erreur a empêché votre désabonnement à la newsletter.");
-                }
-            }
-        }
-    }
-    
-    /**
-     * Delete user account by user
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function updateAccountDelete(Parameter $post)
-    {
-        if($post->get('delete')) {
-            if(!$post->get('deleteConfirm')) {
-                $this->alert->addError("Vous n'avez pas confirmé le souhait de supprimer votre compte.");
-            } else {
-                if($this->validation->validateInput('user', $post)) {
-                    $id = $this->session->get('id');
-                    $passwordHash = $this->userDAO->getUser($id)->getPassword();
-                    if(!password_verify($post->get('password'), $passwordHash)) {
-                        $this->alert->addError("Votre mot de passe n'est pas bon");
-                    } else {
-                        $this->userDAO->deleteUser($id);
-                        if($this->userDAO->existsUser($id)) {
-                            $this->alert->addError("Votre compte n'a pas pu être supprimé.");
-                        } else {
-                            $this->logout();
-                        }
-                    }
-                }
-            }
-        }
-    }
-    
-    /**
      * Display users for administration
      *
      * @param  Parameter $post
@@ -232,99 +113,19 @@ class BackController extends Controller
         }
         if(!$post->get('submit') OR $post->get('delete')) {
             if($post->get('delete')) {
-                $this->deleteUser($post);
+                $this->userController->deleteUser($post);
             }
             return $this->view->render($this->controller, 'users', [
                 'users' => $this->userDAO->getUsers()
             ]);
         }
         if($post->get('update')) {
-            $this->updateUser($post);
+            $this->userController->updateUser($post);
         }
         $user = $this->userDAO->getUser($post->get('id'));
         return $this->view->render($this->controller, 'user', [
             'user' => $user
         ]);
-    }
-    
-    /**
-     * Delete user by administrator
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function deleteUser(Parameter $post)
-    {
-        if($post->get('delete')) {
-            if(!$post->get('deleteConfirm')) {
-                $this->alert->addError("Vous n'avez pas confirmé la suppression de l'utilisateur.");
-            } else {
-                $id = $post->get('id');
-                $this->userDAO->deleteUser($id);
-                if($this->userDAO->existsUser($id)) {
-                    $this->alert->addError("L'utilisateur n'a pas pu être supprimé.");
-                } else {  
-                    $this->alert->addSuccess("L'utilisateur a bien été supprimé.");
-                }
-            }
-        }
-    }
-    
-    /**
-     * Update user by administrator
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function updateUser(Parameter $post)
-    {
-        $id = $post->get('id');
-        $post->delete(['id', 'submit']);
-        $attributesArray = [
-            'pseudo',  
-            'email', 
-            'filename', 
-            'newsletter', 
-            'flag', 
-            'banned', 
-            'role_id'
-        ];
-        if($this->validation->validateInput('user', $post)) {                       
-            foreach($attributesArray as $index => $attribute) {
-                $value = $post->get($attribute);
-                if(!$this->userDAO->updateUser($id, $attribute, $value)) {
-                    $this->alert->addError("Le champ $attribute n'a pas pu être modifié.");
-                }
-            }
-        }
-        $post->set('id', $id);
-    }
-    
-    
-    /**
-     * Add comment by user
-     *
-     * @param  Parameter $post
-     * @return void
-     */
-    public function addComment(Parameter $post)
-    {
-        if(!$this->checkLoggedIn()) {
-            $this->alert->addError("Vous devez être connecté pour laisser un commentaire");
-            header("Location: ".URL."connexion");
-            exit;
-        }
-        if(!$this->validation->validateInput('comment', $post)) {
-            $this->session->set('comment', $post->get('comment'));
-            header("Location:".URL."articles&id=".$post->get('articleId')."#comment");
-            exit;
-        } else {
-            if($this->commentDAO->addComment($post, $this->date)) {
-                $this->alert->addSuccess("Merci, votre commentaire est soumis à validation.");
-                header("Location:".URL."articles");
-                exit;
-            }
-        }
     }
 
 }
