+++ App/src/controller/CategoryController.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\src\controller;
+
+class CategoryController extends Controller
+{
+
+}
