+++ App/src/controller/FrontController.php
@@ -15,6 +15,40 @@ class FrontController extends Controller
      * @var string
      */
     private $controller = 'front';
+
+    /**
+     * Construct FrontController
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        parent::__construct();  
+        $this->articleController = new ArticleController;
+        $this->categoryController = new CategoryController;
+        $this->commentController = new CommentController;
+        $this->userController = new UserController;
+        $this->logController = new LogController;
+    }
+
+    /**
+     * Get login page / Login user
+     *
+     * @param  object $post
+     * @return void|mixed $view
+     */
+    public function login(Parameter $post)
+    {
+        if($this->checkLoggedIn()) {
+            $this->alert->addError("Vous êtes déjà connecté.");
+            ($this->checkAdmin())? header("Location: ".URL."admin") : header("Location: ".URL."profil");
+            exit;
+        }
+        $this->logController->login($post);
+        return $this->view->render($this->controller, 'login', [
+            'post'=> $post
+        ]);
+    }
     
     /**
      * Get homepage
@@ -96,8 +130,11 @@ class FrontController extends Controller
      * @param  int $id
      * @return mixed $view
      */
-    public function single($id)
+    public function single($id, $post = null)
     {
+        if($post->get('submit')) {
+            $this->commentController->addComment($post);
+        }
         $article = Search::lookForOr($this->articleDAO->getArticles(), [
             'id' => $id
         ]);
@@ -106,9 +143,19 @@ class FrontController extends Controller
             header("Location: ".URL."articles");
             exit;
         }
+        $comments = Search::lookForAnd($this->commentDAO->getComments(), [
+            'articleId' => $id,
+            'status' => parent::ACTIVE_COMMENT
+        ]);
+        $users = [];
+        foreach($comments as $comment) {
+            $users[$comment->getUserId()] = $this->userDAO->getUser($comment->getUserId());
+        }
         return $this->view->render($this->controller, 'single', [
             'article' => $article[0],
-            'comment' => ($this->session->get('comment'))? $this->session->show('comment') : ''
+            'users' => $users,
+            'comments' => $comments,
+            'content' => $post->get('content')? $post->get('content') : ''
         ]);
     }
     
@@ -185,47 +232,4 @@ class FrontController extends Controller
         exit;   
     }
     
-    /**
-     * Get login page / Login user
-     *
-     * @param  object $post
-     * @return void|mixed $view
-     */
-    public function login(Parameter $post)
-    {
-        if($this->checkLoggedIn()) {
-            $this->alert->addError("Vous êtes déjà connecté.");
-            ($this->checkAdmin())? header("Location: ".URL."admin") : header("Location: ".URL."profil");
-            exit;
-        }
-        if($post->get('submit')) {
-            if($this->validation->validateInput('user', $post)) {
-                $user = Search::lookForOr($this->userDAO->getUsers(), [
-                    'pseudo' => $post->get('pseudo')
-                ]);
-                if(!$user) {
-                    $this->alert->addError("Vos identifiants sont incorrects.");
-                } else {
-                    if($user[0]->getLevel() < parent::MEMBER_LEVEL) {
-                        $this->alert->addError("Vous devez d'abord valider votre compte.");
-                    } else {
-                        if(!password_verify($post->get('password'), $user[0]->getPassword())) {
-                            $this->alert->addError("Vos identifiants sont incorrects.");
-                        } else {          
-                            $this->userDAO->updateUser($user[0]->getId(), 'last_connection', $this->date);
-                            $this->alert->addSuccess("Content de vous revoir.");
-                            $this->session->set('id', $user[0]->getId());
-                            $this->session->set('level', $user[0]->getLevel());
-                            $this->session->set('pseudo', $user[0]->getPseudo());
-                            ($this->checkAdmin())? header("Location: ".URL."admin") : header("Location: ".URL."profil");
-                            exit;
-                        }
-                    }
-                }
-            }
-        }
-        return $this->view->render($this->controller, 'login', [
-            'post'=> $post
-        ]);
-    }
 }
